feat: table component and rendering
