+++ src/pages/Admin.jsx
@@ -1,235 +1,50 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import React, { useState } from "react";
-import Box from '@mui/joy/Box';
-import Button from '@mui/joy/Button';
-import Divider from '@mui/joy/Divider';
-import FormControl from '@mui/joy/FormControl';
-import FormLabel from '@mui/joy/FormLabel';
-import Link from '@mui/joy/Link';
-import Input from '@mui/joy/Input';
-import Modal from '@mui/joy/Modal';
-import ModalDialog from '@mui/joy/ModalDialog';
-import ModalClose from '@mui/joy/ModalClose';
-import Select from '@mui/joy/Select';
-import Option from '@mui/joy/Option';
-import Table from '@mui/joy/Table';
-import Sheet from '@mui/joy/Sheet';
-import Checkbox from '@mui/joy/Checkbox';
-import IconButton, { iconButtonClasses } from '@mui/joy/IconButton';
-import Typography from '@mui/joy/Typography';
-import Menu from '@mui/joy/Menu';
-import MenuButton from '@mui/joy/MenuButton';
-import MenuItem from '@mui/joy/MenuItem';
-import Dropdown from '@mui/joy/Dropdown';
+import { Box, Typography } from '@mui/joy';
 
-import FilterAltIcon from '@mui/icons-material/FilterAlt';
-import SearchIcon from '@mui/icons-material/Search';
-import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
-import MoreHorizRoundedIcon from '@mui/icons-material/MoreHorizRounded';
+import AdminTable from "../components/AdminTable";
 
-const rows = [
-    {
-        id: '11',
-        date: 'Feb 3, 2023',
-        status: 'approved',
-        content: 'Beaty if pretty cool!',
-        user: {
-            name: 'Jordan Sheehan',
-            email: 'js@ufl.edu',
-        },
-    },
-    {
-        id: '12',
-        date: 'Nov 12, 2024',
-        status: 'inReview',
-        content: 'Questionable review',
-        user: {
-            name: 'Evan Robinson',
-            email: 'evan@ufl.edu',
-        },
-    },
-    {
-        id: '13',
-        date: 'Dec 11, 2024',
-        status: 'rejected',
-        content: 'Profanity detected',
-        user: {
-            name: 'Vance Boudreau',
-            email: 'vboudreau1@ufl.edu',
-        },
-    },
-];
-
-function RowMenu() {
-    return (
-        <Dropdown>
-            <MenuButton
-                slots={{ root: IconButton }}
-                slotProps={{ root: { variant: 'plain', color: 'neutral', size: 'sm' } }}
-            >
-                <MoreHorizRoundedIcon />
-            </MenuButton>
-            <Menu size="sm" sx={{ minWidth: 140 }}>
-                <MenuItem>Approve</MenuItem>
-                <MenuItem>Reject</MenuItem>
-                <Divider />
-                <MenuItem color="danger">Delete</MenuItem>
-            </Menu>
-        </Dropdown>
-    );
-}
 const Admin = () => {
-    const [review, setReview] = useState("");
-    const [selected, setSelected] = useState([]);
-    const [open, setOpen] = useState(false);
-
-    const renderFilters = () => (
-        <React.Fragment>
-            <FormControl size="sm">
-                <FormLabel>Status</FormLabel>
-                <Select
-                    size="sm"
-                    placeholder="Filter by status"
-                    slotProps={{ button: { sx: { whiteSpace: 'nowrap' } } }}
-                >
-                    <Option value="approved">Approved</Option>
-                    <Option value="inReview">In Review</Option>
-                    <Option value="rejected">Rejected</Option>
-                </Select>
-            </FormControl>
-        </React.Fragment>
-    );
-
     return (
-        <React.Fragment>
-            <Sheet
-                className="SearchAndFilters-mobile"
-                sx={{ display: { xs: 'flex', sm: 'none' }, my: 1, gap: 1 }}
-            >
-                <Input
-                    size="sm"
-                    placeholder="Search"
-                    startDecorator={<SearchIcon />}
-                    sx={{ flexGrow: 1 }}
-                />
-                <IconButton
-                    size="sm"
-                    variant="outlined"
-                    color="neutral"
-                    onClick={() => setOpen(true)}
-                >
-                    <FilterAltIcon />
-                </IconButton>
-                <Modal open={open} onClose={() => setOpen(false)}>
-                    <ModalDialog aria-labelledby="filter-modal" layout="fullscreen">
-                        <ModalClose />
-                        <Typography id="filter-modal" level="h2">
-                            Filters
-                        </Typography>
-                        <Divider sx={{ my: 2 }} />
-                        <Sheet sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                            {renderFilters()}
-                            <Button color="primary" onClick={() => setOpen(false)}>
-                                Submit
-                            </Button>
-                        </Sheet>
-                    </ModalDialog>
-                </Modal>
-            </Sheet>
+        <Box sx={{ display: 'flex', minHeight: '100dvh' }}>
             <Box
-                className="SearchAndFilters-tabletUp"
+                component="main"
+                className="MainContent"
                 sx={{
-                    borderRadius: 'sm',
-                    py: 2,
-                    display: { xs: 'none', sm: 'flex' },
-                    flexWrap: 'wrap',
-                    gap: 1.5,
-                    '& > *': {
-                        minWidth: { xs: '120px', md: '160px' },
+                    px: { xs: 2, md: 6 },
+                    pt: {
+                        xs: 'calc(12px + var(--Header-height))',
+                        sm: 'calc(12px + var(--Header-height))',
+                        md: 3,
                     },
+                    pb: { xs: 2, sm: 2, md: 3 },
+                    flex: 1,
+                    display: 'flex',
+                    flexDirection: 'column',
+                    minWidth: 0,
+                    height: '100dvh',
+                    gap: 1,
                 }}
             >
-                {renderFilters()}
-            </Box>
-            <Sheet
-                className="OrderTableContainer"
-                variant="outlined"
-                sx={{
-                    display: { xs: 'none', sm: 'initial' },
-                    width: '100%',
-                    borderRadius: 'sm',
-                    flexShrink: 1,
-                    overflow: 'auto',
-                    minHeight: 0,
-                }}
-            >
-                <Table
-                    aria-labelledby="tableTitle"
-                    stickyHeader
-                    hoverRow
+                <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                </Box>
+                <Box
                     sx={{
-                        '--TableCell-headBackground': 'var(--joy-palette-background-level1)',
-                        '--Table-headerUnderlineThickness': '1px',
-                        '--TableRow-hoverBackground': 'var(--joy-palette-background-level1)',
-                        '--TableCell-paddingY': '4px',
-                        '--TableCell-paddingX': '8px',
+                        display: 'flex',
+                        mb: 1,
+                        gap: 1,
+                        flexDirection: { xs: 'column', sm: 'row' },
+                        alignItems: { xs: 'start', sm: 'center' },
+                        flexWrap: 'wrap',
+                        justifyContent: 'space-between',
                     }}
                 >
-                    <thead>
-                        <tr>
-                            <th style={{ width: 48, textAlign: 'center', padding: '12px 6px' }}>
-                                <Checkbox
-                                    size="sm"
-                                    indeterminate={
-                                        selected.length > 0 && selected.length !== rows.length
-                                    }
-                                    checked={selected.length === rows.length}
-                                    onChange={(event) => {
-                                        setSelected(
-                                            event.target.checked ? rows.map((row) => row.id) : [],
-                                        );
-                                    }}
-                                    color={
-                                        selected.length > 0 || selected.length === rows.length
-                                            ? 'primary'
-                                            : undefined
-                                    }
-                                    sx={{ verticalAlign: 'text-bottom' }}
-                                />
-                            </th>
-                            <th style={{ width: 120, padding: '12px 6px' }}>
-                                <Link
-                                    underline="none"
-                                    color="primary"
-                                    component="button"
-                                    onClick={() => { }}
-                                    endDecorator={<ArrowDropDownIcon />}
-                                >
-                                    Review
-                                </Link>
-                            </th>
-                            <th style={{ width: 140, padding: '12px 6px' }}>Date</th>
-                            <th style={{ width: 140, padding: '12px 6px' }}>Status</th>
-                            <th style={{ width: 240, padding: '12px 6px' }}>Content</th>
-                            <th style={{ width: 240, padding: '12px 6px' }}>User</th>
-                        </tr>
-                    </thead>
-                </Table>
-            </Sheet>
-            <Box
-                className="Pagination-laptopUp"
-                sx={{
-                    pt: 2,
-                    gap: 1,
-                    [`& .${iconButtonClasses.root}`]: { borderRadius: '50%' },
-                    display: {
-                        xs: 'none',
-                        md: 'flex',
-                    },
-                }}
-            >
+                    <Typography level="h2" component="h1">
+                        Reviews
+                    </Typography>
+                </Box>
+                <AdminTable />
             </Box>
-        </React.Fragment>
+        </Box>
     );
 }
